+++ rust/operator-binary/src/controller.rs
@@ -381,10 +381,7 @@ impl OpaClusterConfigFile {
             decision_logs: decision_logging,
             // Enable more Prometheus metrics, such as bundle loads
             // See https://www.openpolicyagent.org/docs/monitoring#status-metrics
-            status: Some(OpaClusterConfigStatus {
-                service: OPA_STACKABLE_SERVICE_NAME.to_owned(),
-                prometheus: true,
-            }),
+            status: Some(OpaClusterConfigStatus { prometheus: true }),
         }
     }
 }
@@ -421,7 +418,6 @@ pub struct OpaClusterConfigDecisionLog {
 
 #[derive(Serialize, Deserialize)]
 struct OpaClusterConfigStatus {
-    service: String,
     prometheus: bool,
 }